keystone: refactor fernet key rotation timing
Previously this relied on hardcoded rotation times from hiera;
it also had the timing exactly backwards so that new
keys were immediately overwritten via an rsync call.

The new plan uses division to distribute key rotations
around the clock, and organizes the rsyncs so that
each change is propogated to the other keystone hosts
after a key rotation.

In addition to removing some hiera, this has the bonus
advantage of preventing simultaneous rsyncs to any given
host which should avoid collision errors.

Bug: T317838

+++ modules/profile/manifests/openstack/base/keystone/fernet_keys.pp
@@ -1,20 +1,6 @@
 class profile::openstack::base::keystone::fernet_keys(
     Array[Stdlib::Fqdn] $keystone_hosts = lookup('profile::openstack::base::openstack_controllers'),
-    String $rotate_time = lookup('profile::openstack::base::rotate_time'),
-    String $sync_time = lookup('profile::openstack::base::sync_time'),
     ) {
-
-    systemd::timer::job { 'keystone_rotate_keys':
-        description        => 'Rotate keys for Keystone fernet tokens',
-        command            => '/usr/bin/keystone-manage fernet_rotate --keystone-user keystone --keystone-group keystone',
-        interval           => {
-            'start'    => 'OnCalendar',
-            'interval' => "*-*-* ${rotate_time}",
-        },
-        logging_enabled    => true,
-        user               => 'root',
-        monitoring_enabled => false,
-    }
 
     file { '/etc/keystone/fernet-keys':
         ensure => directory,
@@ -32,18 +18,43 @@ class profile::openstack::base::keystone::fernet_keys(
         read_only   => true,
     }
 
-    $other_hosts = $keystone_hosts - $::fqdn
-    $other_hosts.each |String $thishost| {
-        systemd::timer::job { "keystone_sync_keys_to_${thishost}":
-            description        => "Sync keys for Keystone fernet tokens to ${thishost}",
-            command            => "/usr/bin/rsync -a --delete rsync://${thishost}/keystonefernetkeys/ /etc/keystone/fernet-keys/",
-            interval           => {
-            'start'    => 'OnCalendar',
-            'interval' => "*-*-* ${sync_time}",
-            },
-            logging_enabled    => true,
-            monitoring_enabled => false,
-            user               => 'keystone',
+    # It's important to do these steps in the right order: a host should rotate its keys, and immediately
+    #  after that each other host should rsync to pick up the changes.
+    # Rotations happen on the hour, and syncing on the half.
+    #
+    # Note that if the order of hosts in $keystone_hosts is not consistent across all
+    #  hosts this will cause chaos.
+    #
+    $hostcount = count($keystone_hosts)
+    $staggerhours = 24/$hostcount
+
+    $keystone_hosts.each |$index, String $thishost| {
+        $activehour = $index * $staggerhours
+
+        if $thishost == $::fqdn {
+            systemd::timer::job { 'keystone_rotate_keys':
+                description        => 'Rotate keys for Keystone fernet tokens',
+                command            => '/usr/bin/keystone-manage fernet_rotate --keystone-user keystone --keystone-group keystone',
+                interval           => {
+                    'start'    => 'OnCalendar',
+                    'interval' => "*-*-* ${activehour}:00:00",
+                },
+                logging_enabled    => true,
+                user               => 'root',
+                monitoring_enabled => false,
+            }
+        } else {
+            systemd::timer::job { "keystone_sync_keys_from_${thishost}":
+                description        => "Sync keys for Keystone fernet tokens to ${thishost}",
+                command            => "/usr/bin/rsync -a --delete rsync://${thishost}/keystonefernetkeys/ /etc/keystone/fernet-keys/",
+                interval           => {
+                'start'    => 'OnCalendar',
+                'interval' => "*-*-* ${activehour}:30:00",
+                },
+                logging_enabled    => true,
+                monitoring_enabled => false,
+                user               => 'keystone',
+            }
         }
     }
 }
